fix: store inline bigint ranges and pin SQLite forUpdate rows
Inline collection create remaps biginteger to bigint before the numeric-range branch, so bigint columns were stored without formatOptions. The public float type is double, which made the inline numeric E2E 400. Pin database so SQLite forUpdate rows no longer leak PDO column indexes into versioned updates.

+++ tests/e2e/Services/Databases/DatabasesBase.php
@@ -1598,7 +1598,7 @@ trait DatabasesBase
             $schemaResource => [
                 ['key' => 'counter', 'type' => ColumnType::Integer->value],
                 ['key' => 'total', 'type' => ColumnType::BigInteger->value],
-                ['key' => 'ratio', 'type' => ColumnType::Float->value],
+                ['key' => 'ratio', 'type' => ColumnType::Double->value],
                 ['key' => 'bounded', 'type' => ColumnType::Integer->value, 'min' => 0, 'max' => 100],
                 // The numeric endpoints take no size, so this one cannot narrow the
                 // column below the int64 range the definition still declares
